requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudinary==1.21.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==19.9.0
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.3
Pillow==7.1.2
psycopg2==2.8.3
pycodestyle==2.6.0
pycparser==2.20
# pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==4.1.4
